// frontend/src/app/api/models/server-config.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Ollama's defaults for models without own values. From server.log and/or [modules.catalog].
 */
export interface ServerConfig {
  contextLength?: (number | null);
  flashAttention?: (boolean | null);
  keepAlive?: (string | null);
  kvCacheType?: (string | null);
  logFile?: (string | null);
  maxLoadedModels?: (number | null);
  note?: (string | null);
  numParallel?: (number | null);
  overridden?: Array<string>;
  source: 'log' | 'config' | 'mixed' | 'unknown';
}